controller/requirements/base.txt
# SHA1:7c28a44e674477846c9a28bc6bd27676ee44d9d4
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
certifi==2026.2.25
    # via requests
cffi==1.17.1
    # via hidapi-usb
charset-normalizer==3.4.6
    # via requests
googleapis-common-protos==1.73.1
    # via opentelemetry-exporter-otlp-proto-http
hidapi-usb==0.3.1
    # via pydualsense
idna==3.11
    # via requests
importlib-metadata==8.7.1
    # via opentelemetry-api
numpy==2.2.1
    # via
    #   -r requirements/base.in
    #   opencv-python-headless
opencv-python-headless==4.10.0.84
    # via -r requirements/base.in
opentelemetry-api==1.40.0
    # via
    #   -r requirements/base.in
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp-proto-common==1.40.0
    # via opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-http==1.40.0
    # via -r requirements/base.in
opentelemetry-proto==1.40.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.40.0
    # via
    #   -r requirements/base.in
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.61b0
    # via opentelemetry-sdk
protobuf==6.33.6
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
pycparser==2.22
    # via cffi
pydualsense==0.7.1
    # via -r requirements/base.in
requests==2.33.1
    # via opentelemetry-exporter-otlp-proto-http
typing-extensions==4.15.0
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
urllib3==2.6.3
    # via requests
websockets==14.1
    # via -r requirements/base.in
zipp==3.23.0
    # via importlib-metadata
